End Time Calculation Adjustments
Because this is in .chart format Hero Power should last for 7 seconds, but the value of beatsPerSecond is calculated wrong. Change "6000" to "210000" so the value can be scaled down to how it is in the chart file. New method for half notes it checks if the note is equal distance apart from the last note and the next note. The time from this note to the next note is added on to the end time. This is still not a perfect calculation and thinking on implementing this with another format other than chart or mid, XMK.